third_party/blink/web_tests/external/wpt/custom-elements/registries/ShadowRoot-init-declarative-expected.txt

This is a testharness.js-based test.
[FAIL] Custom element inside 'shadowrootcustomelementregistry' declarative shadow root
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Built-in element inside 'shadowrootcustomelementregistry' declarative shadow root
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
